Fix sorting stackable equippables.
As of DST build 218082, GiveItem will crash the game if the slot to be used is
one of the equipslots. As a quick fix, the sorting code now watches out for
cases when GetNextAvailableInventorySlot returns an equipslot, and manually
handles topping up the equipped stack, and putting the leftovers to an itemslot.